backend-python/requirements.txt
Flask
Pillow
Flask-Cors
bson